// src/shiphero/generated/operations.ts
export interface ProductQueryArgs {
  /**
   * Filter by ID; pass the public API ID unless the field name says legacy ID.
   * GraphQL type: `String`
   */

  id?: string | null;
  /**
   * Filter by SKU. SKU matching is case-insensitive and ignores surrounding whitespace.
   * GraphQL type: `String`
   */

  sku?: string | null;
  /**
   * Filter by barcode.
   * GraphQL type: `String`
   */

  barcode?: string | null;
  /**
   * If you are a 3PL you can specify one of your customer's account.
   * GraphQL type: `String`
   */

  customer_account_id?: string | null;
  /**
   * If true, the query will only be analyzed (cost estimated) without executing it.
   * GraphQL type: `Boolean`
   */

  analyze?: boolean | null;
}

/**
 * Fetch products matching the supplied arguments.
 * Arguments for the `products` query.
 */

export interface ProductsQueryArgs {
  /**
   * Filter by SKU. SKU matching is case-insensitive and ignores surrounding whitespace.
   * GraphQL type: `String`
   */

  sku?: string | null;
  /**
   * Inclusive start of the created range, in ISO-8601 datetime format.
   * GraphQL type: `ISODateTime`
   */

  created_from?: string | null;
  /**
   * Inclusive end of the created range, in ISO-8601 datetime format.
   * GraphQL type: `ISODateTime`
   */

  created_to?: string | null;
  /**
   * Inclusive start of the updated range, in ISO-8601 datetime format.
   * GraphQL type: `ISODateTime`
   */

  updated_from?: string | null;
  /**
   * Inclusive end of the updated range, in ISO-8601 datetime format.
   * GraphQL type: `ISODateTime`
   */

  updated_to?: string | null;
  /**
   * If you are a 3PL you can specify one of your customer's account.
   * GraphQL type: `String`
   */

  customer_account_id?: string | null;
  /**
   * When true, filters results where has kits applies; when false, filters where it does not.
   * GraphQL type: `Boolean`
   */

  has_kits?: boolean | null;
  /**
   * Filters products that have ANY of the specified tags (OR).
   * GraphQL type: `[String]`
   */

  tag?: string[] | null;
  /**
   * Excludes products that have ANY of the specified tags; untagged products are included.
   * GraphQL type: `[String]`
   */

  tag_not_in?: string[] | null;
  /**
   * If true, the query will only be analyzed (cost estimated) without executing it.
   * GraphQL type: `Boolean`
   */

  analyze?: boolean | null;
}